server/requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ephem==4.1.5
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
Flask-MQTT==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.2
six==1.16.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
